Implemented simple echo effect and parsing tokens
